api/requirements.txt
qrcode
pillow
jinja2
html2image
fastapi
uvicorn
pydantic